LayoutTests/imported/w3c/web-platform-tests/payment-request/payment-request-abort-method.https-expected.txt

CONSOLE MESSAGE: Unhandled Promise Rejection: InvalidStateError: The object is in an invalid state.
CONSOLE MESSAGE: Unhandled Promise Rejection: InvalidStateError: The object is in an invalid state.
CONSOLE MESSAGE: Unhandled Promise Rejection: InvalidStateError: The object is in an invalid state.
This test requires user interaction.
Please click here to allow show payment request. If you find a buggy test, please file a bug and tag one of the suggested reviewers.

Harness Error (TIMEOUT), message = null

PASS Throws if the promise [[state]] is not "interactive"
PASS Calling abort() multiple times is always a new object.
TIMEOUT The same request cannot be shown multiple times. Test timed out
NOTRUN Aborting a request before it is shown doesn't prevent it from being shown later.
